+++ tests/conf_json_test.cc
@@ -1,4 +1,5 @@
 ﻿#include "gtest/gtest.h"
+#include "tkc/named_value.h"
 #include "conf_io/conf_json.h"
 
 TEST(ConfJson, arr) {
@@ -310,6 +311,48 @@ TEST(Json, create) {
   ASSERT_NE(conf, (tk_object_t*)NULL);
   ASSERT_EQ(tk_object_set_prop_int(conf, "value", 123), RET_OK);
   ASSERT_EQ(tk_object_get_prop_int(conf, "value", 0), 123);
+  TK_OBJECT_UNREF(conf);
+}
+
+static ret_t on_prop(void* ctx, const void* data) {
+  char buff[64];
+  str_t* str = (str_t*)ctx;
+  named_value_t* nv = (named_value_t*)data;
+
+  str_append_more(str, nv->name, "=", value_str_ex(&(nv->value), buff, sizeof(buff)), "\n", NULL);
+
+  return RET_OK;
+}
+
+TEST(Json, foreach) {
+  str_t str;
+  tk_object_t* conf = conf_json_create();
+  ASSERT_NE(conf, (tk_object_t*)NULL);
+  ASSERT_EQ(tk_object_set_prop_str(conf, "name", "awtk"), RET_OK);
+  ASSERT_EQ(tk_object_set_prop_int(conf, "value", 123), RET_OK);
+
+  str_init(&str, 100);
+  ASSERT_EQ(tk_object_foreach_prop(conf, on_prop, &str), RET_OK);
+  ASSERT_STREQ(str.str, "name=awtk\nvalue=123\n");
+
+  str_reset(&str);
+  TK_OBJECT_UNREF(conf);
+}
+
+TEST(Json, foreach1) {
+  str_t str;
+  tk_object_t* conf = conf_json_create();
+  ASSERT_NE(conf, (tk_object_t*)NULL);
+  ASSERT_EQ(tk_object_set_prop_str(conf, "name", "awtk"), RET_OK);
+  ASSERT_EQ(tk_object_set_prop_int(conf, "value", 123), RET_OK);
+  ASSERT_EQ(tk_object_set_prop_int(conf, "detail.age", 123), RET_OK);
+  ASSERT_EQ(tk_object_set_prop_int(conf, "detail.salary", 1000), RET_OK);
+
+  str_init(&str, 100);
+  ASSERT_EQ(tk_object_foreach_prop(conf, on_prop, &str), RET_OK);
+  ASSERT_STREQ(str.str, "name=awtk\nvalue=123\n");
+
+  str_reset(&str);
   TK_OBJECT_UNREF(conf);
 }